Fix compilation error part 2
src/dialogs/gnome-cmd-search-dialog.cc:1023:24: error: this statement may fall through [-Werror=implicit-fallthrough=]
  - gcc7 new waring. Just disabled with pragma, but you must review this (whether fallthrough here is actually intended or not)

src/tags/gnome-cmd-tags-exiv2.cc:71:17: error: format '%u' expects argument of type 'unsigned int', but argument 4 has type 'long int'
src/tags/gnome-cmd-tags-exiv2.cc:71:17: error: format '%u' expects argument of type 'unsigned int', but argument 4 has type 'long int' [-Werror=format=]
  - Should use %ld

src/tags/gnome-cmd-tags-taglib.cc:94:29: error: unused variable 'emphasis_names' [-Werror=unused-variable]
  - This variable is in the comment. For now, disabled warning with pragma

rc/tags/gnome-cmd-tags-taglib.cc:180:16: error: enumeration value 'TAG_AUDIO_ALBUMARTIST' not handled in switch [-Werror=switch-enum]
  - Nuked warning with pragma

src/tags/gnome-cmd-tags-taglib.cc:307:6: error: no previous declaration for 'void gcmd_tags_taglib_init()
  - Add needed header

src/tags/gnome-cmd-tags-taglib.cc:412:18: error: ISO C++ forbids converting a string constant to 'gchar* {
aka char*}'
  - Change struct member to "const gchar *"

src/tags/gnome-cmd-tags-taglib.cc:166:13: error: inlining failed in call to 'void readTags(GnomeCmdFileMetadata&, const TagLib::ID3v2::Tag*)': --param max-inline-insns-single limit reached [-Werror=inline]
src/tags/gnome-cmd-tags-taglib.cc:166:13: error: inlining failed in call to 'void readTags(GnomeCmdFileMetadata&, const TagLib::ID3v2::Tag*)': --param max-inline-insns-single limit reached [-Werror=inline]
  - Function size too long, killed inline

src/intviewer/bm_chartype.cc:52:42: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]
  - Use GINT_TO_POINTER

src/intviewer/datapresentation.cc:107:12: error: switch missing default case [-Werror=switch-default]
  - easy fix

src/intviewer/fileops.cc:239:33: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
src/intviewer/fileops.cc:259:33: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
src/intviewer/fileops.cc:308:36: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  - Just ignored this for now by pragma. "Fixing" this can make more mistakes.

src/intviewer/image-render.cc:829:133: error: 'GThread* g_thread_create(GThreadFunc, gpointer, gboolean, GError**)' is deprecated: Use 'g_thread_new' instead [-Werror=deprecated-declarations]
  - Newly written code should use g_thread_new. Please review this.

src/intviewer/inputmodes.cc:393:59: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
  - Cast INVALID_CHAR with int

src/intviewer/search-dlg.cc:125:5: error: "HEX_HISTORY" is not defined, evaluates to 0 [-Werror=undef]
 #if HEX_HISTORY
  - Change to ifdef

src/intviewer/search-progress-dlg.cc:138:12: error: no previous declaration for 'GtkWidget* gviewer_search_progress_dlg_new(GtkWindow*)' [-Werror=missing-declarations]
  - Change to static

+++ src/intviewer/inputmodes.cc
@@ -390,7 +390,7 @@ inline gboolean utf8_is_valid_char(GVInputModesData *imd, offset_type offset)
 {
     int len = utf8_get_char_len(imd, offset);
 
-    if (len==0 || (gv_input_mode_get_byte(imd, offset+len)==INVALID_CHAR))
+    if (len==0 || (gv_input_mode_get_byte(imd, offset+len)==(int)INVALID_CHAR))
         return FALSE;
 
     if (len==1)
